drivers: misc: pn547: Remove deprecated code and add clk_req gpio entry
The msm_xo_get function likely is dated back to ARM32 tree, remove it along
with its dependents as we are not using them anyways. Also, we configure
clk_req_gpio in k4.4 but because of its dependency to deprecated code, we
don't use it here; add it and properly configure it.

Signed-off-by: ederekun <sedrickvince@gmail.com>