sched/tune: Use an IRQ instead of kthreads
We aim to switch boost on and off as soon as possible. Even though
kthreads do this well, we create virtually unkillable threads that
the scheduler does not like; which is why they're interruptible.
But with our code being minimal and spontaneous due to it being
dependent on external inputs, it is better to utilize an interrupt.

This reduces the time spent from input to completion from 200us-1ms
to 20-200us, making its impact very minimal and the switch from
boosted to non-boosted state almost instantaneous even for the
scheduler.

In addition to this, restrict external access to schedtune input
components by instead allocating a local struct at runtime. Also
eliminate the boost boolean as the timer's pending state already
represents the same data.

Signed-off-by: ederekun <sedrickvince@gmail.com>