clk: qcom: clk-rcg2: Reconfigure byte and pixel clks at recalc
The common clk framework works by propagating clk rate from parent
to children. This behavior did not exist before with the msm clk
framework which manually set clks through set rate, mostly propagating
only upwards. Because of this, CAF forgot to take into consideration
or may have workaround this mechanism as they have not created
functions that reset or reconfigures the registers whenever recalc
is called.

This is very important as recalc sets the clk's rate, and because
of this there can be cases where a recalc can be called first prior
to set rate. Set rate will no longer be called as it assumes that
the recalc function already updated the rate AND registries when the
parent rate changed first; despite the rate the only one being
updated, causing the registries to be not synced.

This fixes the display not being able to come back up after suspend,
eliminating the need for the reconf link hack.

Signed-off-by: ederekun <sedrickvince@gmail.com>