clk: qcom: mmcc-msm8998: Introduce RECONFIGURE_RCG flag for multimedia clocks
Our platform msm8998/msmcobalt has a HW limitation that causes MM RCGs to
end up running at XO frequency after multimedia power collapse. Because
of this, downstream configured CLKFLAG_NO_RATE_CACHE on all MM clocks
to have the MSM clock framework set the rate and parent always whenever
the driver requests it.

Downstream commit:
https://github.com/LineageOS/android_kernel_oneplus_msm8998/commit/52168e4c325387c27ace51956752462f0898b2b8

With the common clock framework, however, this is not possible. There
are no flags that disable this optimization, and CLK_GET_RATE_NOCACHE
only enables a path that only gets triggered whenever a driver calls
on the clk_get_rate function (something that is seldom used in any of
our drivers) and that it does not even update the configuration of the
clocks, just updates the cached rate in the clock framework - hence,
it is useless when it comes to solving our problem. The parent has no
way of getting reconfigured as well.

Hence, let us introduce instead a prepare routine to be enabled by a
flag that will set the last cached parent and rate in order to perform
the reconfiguration directly at prepare before being enabled - as stated
by the commit above.

This resolves all performance issues with multimedia clocks.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>