fbdev: mdss: Migrate DSI panel modes and HBM to livedisplay driver
Instead of having separate old and new nodes for the panel modes,
we can simply use the livedisplay's streamlined code to register
panel mode cmds, then create functions to wrap the preset function
of the livedisplay for the old nodes to use and send cmds to the
panel. This allows us to keep compatibility with systems that use
the old nodes while also supporting livedisplay for LOS-based
ROMs that may need it. This also accomodates for kernels that have
defined cmds based on the properties provided by the driver, such
as LOS, instead of the mdss-dsi-panel props.

In addition to this, apply the limitations set to the DSI panel
modes to be explicitly enforced in livedisplay and set the default
panel mode to sRGB.

Remove unused panel modes that do not have cmds defined.

Signed-off-by: ederekun <sedrickvince@gmail.com>