irqchip: qcom: mpm: Improve MPM LPM loop
Separate the initialization loop from the vote loop and centralize
all calls to vote_lpm_clk to be only done by the work loop so that
the mutex lock can be removed. Also remove the unnecessary static
definition of the created workqueue_struct as we need the address
only once at init.

Suspend the work loop whenever MPM enters suspend as we only need
to check and vote for LPM once when switching from suspend to
wake. Make this interruptible as to not contribute to system load
and apply this to the polling path as well. Wake up the work loop
everytime msm_mpm_in_suspend changes.

Remove broken clk enable XO fallback when workqueue_struct doesn't
get created and instead use the API provided long wq just in case.

Also simplify the logs.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>