irqchip: qcom: mpm: Introduce MPM LPM based on tracked unlisted IRQs
This has been forwardported from k4.4. This essentially allows MPM
to vote on the XO clock based on the presence of unlisted IRQs and
the device's lpm state. This is to allow the system to detect
and track non-monitorable interrupts as the XO shuts down; in this
case, we vote on XO temporarily while those interrupts are enabled.

We also implement the set_wake logic as to distinguish wake state
from slumber state, and to only base the decision to vote on XO
depending on which IRQs have been enabled to function in each state.
Only touch unlisted_irqs when doing set_wake, do not do anything
with the registers as the wake capability is determined by
firmware/hardware.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>