clk: qcom: gcc-msm8998: Remove postdiv clks and vco_table
Configure these clks according to k4.4, the PLL clk is able to
provide the rate without problems and the way it has been configured
in k4.4 is that the output clk returns the exact fixed clock rate
provided by the PLL clock - essentially a passthrough. So instead
of complicating things, just remove the postdiv clks and name
the main PLL clks as out_main.

There are also no registered clocks in k4.4 that shows the existence
of a USER_CTL configuration for GPLL clks which is what post_div
clocks base their div value from.

Also remove the vco_table as they are never configured in k4.4
and is not necessary as our PLL clks are fixed and determined by
register configs (vco table is used to assist in determining the
rate as set by child clocks).

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>