clk: qcom: gcc-msm8998: Remove unused postdiv clks
Configure these clks according to k4.4, the PLL clk is able to
provide the rate without problems and the way it has been configured
in k4.4 is that the output clk returns the exact fixed clock rate
provided by the PLL clock - essentially a passthrough. So instead
of complicating things, just remove the postdiv clks and name
the main PLL clks as out_main. The clks are unused anyways by
linux.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>