clk: qcom: gpucc-msm8998: Directly connect gfx3d to gpu_pll0
There is no need for the postdiv clock as it is always set to
div-1 which essentially means passthrough towards the main Pll0
clock. What postdiv clock is supposed to do is configure
the register to 0x1 at the USER_CTL address and is supposed to
return a halved rate.

The current setup does NOT match the registers in k4.4 nor
does it match the intended setup.

As shown in tests, the postdiv clock is not functioning correctly
as it needs to pass a rate towards the PLL0 which ends up syncing
to gfx3d due to the common clock framework following the pre-div=1
config at gfx3d and the SET_RATE_PARENT flag. This leads
to a 1:1 rate that follows the rate of gfx3d which is half the
PLL's optimal clock speeds. We use the CRC div for this, which
also sets the postdiv to div-1 and makes the postdiv as a passthrough
clock. The clk is not configured towards a HW clock and only serves
as a postdiv config clk for the PLL0, so it is safe to remove.

In this case, the postdiv is not necessary and could be removed,
and we could instead write 0x1 (div-2) to USER_CTL as to match
k4.4 regs for the PLL clock. This will make it a lot faster for the
rate requests from the gfx3d clk to reach the PLL clk.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>