power: supply: qcom: Extend wait period for dash to transition
There can be instances when the user plugs the phone at a battery
percentage that is considered full for fast charge leading to
the IOCTL being invoked while the charger is still disabled (due
to ongoing init).

Instead of checking once at notify dash unplug, perform this
task on a work thread to allow us to wait for 2 seconds before
checking again the charging status, this should be enough time for
dash to initialize and settle.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>