irqchip: qcom: mpm: Have unlisted IRQs directly vote on the XO clk
Instead of polling, have the IRQs wake the loop whenever they set
enable or wake. This essentially allows the IRQs to vote on the XO
clock directly in real-time, significantly reducing detection
latency as they are only acknowledged when XO is on. With this,
it is also necessary that we greatly optimize the loop and make the
shared variables atomic as the loop is expected to become really
hot whenever interrupt storms occur.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>