irqchip: qcom: mpm: Evaluate LPM votes at unlisted IRQ toggle
Instead of waking the loop everytime there is a bit toggled in
the bitmap and have the loop check the bitmap, have the bitmap
checked as soon as the bit is toggled by the IRQ then evaluate
the LPM vote of the IRQ's group whether it has changed or not.

This is to further reduce unnecessary TTWU attempts on the MPM
thread by the IRQs, making the loop significantly more efficient.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>