irqchip: qcom: mpm: Use IRQF_NO_SUSPEND and pm_system_wakeup()
Instead of using enable_irq_wake() which does not immediately wake
the system at IRQ trigger and of which has also been advised
against its usage with IRQF_NO_SUSPEND [1], use pm_system_wakeup()
at the interrupt function instead. This is inline with the patch
on qcom_glink_native given the fact the driver requires both to
be set [2].

This will ensure that the driver is able to operate at suspend
while also being able to wakeup the system when needed, without
going against the IRQ framework's restriction.

In addition to this, the patch has been adjusted to MPM wherein
we use the system_pm_ops instead as an indicator of whenever the
system enters deep sleep, as the interrupt is supposed to be
triggered after a system wakeup and that RPM must have already
exited sleep or been unmasked. This is because the MPM interrupt
clears all pending interrupts of which are only ran once the system
wakes up, otherwise they get stuck at pending and are likely to
be missed once the interrupt gets triggered again since it got
cleared.

[1]: https://github.com/ederevx/x-ft_kernel_oneplus_msm8998/blob/7438b633a6b073d66a3fa3678ec0dd5928caa4af/Documentation/power/suspend-and-interrupts.txt
[2]: https://github.com/ederevx/x-ft_kernel_oneplus_msm8998/commit/8dad6a832262633adacba4edc137a920dead872e

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>