irqchip: qcom: mpm: Implement set_wake logic from mpm-of
This is based on the implementation in k4.4 mpm-of driver wherein
the enable register is updated based on the from_idle boolean
after writing to the MPM timer. This will have the MPM driver
handle wake interrupts the same way as it was done in mpm-of which
is better fitted for our SoCs' configuration, allowing these
interrupts to function properly or at least similarly to how they
did downstream.

This is to fix the black screen issue, or the device unable to no
longer continue from suspend unless a hard reset is performed.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>