sched/tune: Have schedtune boostgroup send SCHED_CPUFREQ_DL at timeout
When the boostgroup has been in timeout and a boost update has been
received that increases boost_max, send SCHED_CPUFREQ_DL to cpufreq
to poke the governor and have it set the CPU frequency of the
boostgroup to max for a duration equivalent to down_rate_delay_ns.
This also clears the cached next freq so the CPU will decide on the
next frequency based on the applied boost.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>