clk: qcom: gpucc-msm8998: Directly connect the GFX clock to the PLL
The GFX CRC sequence was not ported to mainline, nor is it in use
in sdm660. This sequence causes the clock rate of the GFX clock to
be half the PLL in HW, which is the reason why when setting 2 to
the GFX's pre_div with the CRC sequence, it causes the performance
to be halved. Remove this and finally return to the even PLL post
div with the pre_div configured to 2 as it now performs as
intended.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>