irqchip: qcom: mpm: Restore the enable register at exit sleep
With the set_wake functionality ported to the MPM driver, we are
now able to configure wakeup and enable IRQs properly on the
MPM's enable register. However, the set_wake function is only
called whenever we enter deep sleep or CPU idle, but not when we
exit. It is important that we switch back to enable IRQs after
sleep (deep sleep or CPU idle). Otherwise, MPM will not catch these
IRQs whenever we resume and will not be able to stop the CPU from
idling during runtime - causing freezes. [1]

[1]: https://patchwork.kernel.org/project/linux-arm-msm/patch/1585586460-3272-2-git-send-email-mkshah@codeaurora.org/

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>