power: qcom: smb-lib: Use battery/charging current to determine usage of workaround
There are cases when dash remains to be on while the battery
continues to discharge due to very low current being passed through
towards the battery, despite the power supply indicating that it
is charging. Instead of checking dash_on status, fetch the current
and switch to normal charging if it falls below 500mA which is the
maximum current for USB2.0 and is terribly slow for DCP or dedicated
charging port chargers.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>