irqchip: qcom: mpm: Switch to enable IRQs after setting pending IRQs
Performing the switch to enable IRQs at system ops exit may cause
the MPM IRQ handler to no longer see the wake IRQs if the exit
happens before the IRQ gets handled. Move this call in the
handler after setting pending IRQs to make sure that this does
not happen.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>