clk: qcom: gcc-msm8998: Apply USB30 master branch clock's .depends property
There was a .depends property that was set on the USB30 master
clock, unfortunately this feature did not carry on k4.14. To work
around this lack of support, let's set the AXI clock as USB30 master
clock's parent to have it be enabled by the master branch clock
and have the branch AXI clock pass the rates from the branch clock
towards the USB30 master source clock. This should be fine as both
branch clocks have the same source clock and are likely using the
same rates.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>