power: qcom: smb-lib: Use battery SOC or percentage instead of current
The charging current is unreliable when determining charging state.
Monitor the battery SOC instead to ensure that the battery is charging
after checking battery power supply. Also increase wait duration in
order to accomodate for possible slow MCU response if plugged at an
SOC that is considered dash full to avoid false positives (not charging).

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>