power: qcom: op_cg_uovp: Remove UVP configuration and simplify voting
This is causing unexpected behaviors and is better not done. Also
get client vote instead of the effective_result to avoid endless looping
whenever the votes do not register for some reason. Rerun AICL
instead of calling power_supply_changed as to better account for the
ICL changes for various power supplies.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>