FIXUP: BACKPORT: sched/uclamp: Fix fits_capacity() check in feec()
The changes were not carried over completely and had missed the
main point of the patch. According to the patch's lore, this
changes the fits_capacity behavior, wherein we max aggregate both
the task and the RQs clamp values if the RQ is idle. In k4.14,
the fits_capacity is similar if not identical with the
task_fits_capacity function, and so let's apply the logic to that
function [1].

References:
[1]: https://lore.kernel.org/all/20220804143609.515789-4-qais.yousef@arm.com/

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>