sched/fair: Apply UCLAMP and runnable boosting to compute_energy
After cross examining the 4.14 and 4.19 scheduler, the utilization
calculation can be matched one to one - wherein the busy time
computation in 4.19 refers to the group util normalization, whereas
the performance domain refers to the calculation of the group max
util.

With this, apply clamping to each utilization calculation accordingly
and also perform runnable boosting to max util calculation as seen
in k4.19.

References:
[1]: https://github.com/Juansheng/kernel_oneplus_sm8250/commit/202b305892194de7b1ab562ad85e4ac5e901a9dc#diff-75a417d0e6ebb27175a48c6810294ab6e21847776b1e46c62de9cccfb428883eR6523
[2]: https://github.com/Juansheng/kernel_oneplus_sm8250/commit/9a843ff48d343b24139c7edb7eac92500e733d0a

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>