BACKPORT: sched/uclamp: Make select_idle_capacity() use util_fits_cpu()
This is a modified version of:

https://github.com/Juansheng/kernel_oneplus_sm8250/commit/7211a00a8373340b334516119ca3f7b87a962f09#diff-75a417d0e6ebb27175a48c6810294ab6e21847776b1e46c62de9cccfb428883eR6891

In this patch we instead patch the select_idle_sibling_cstate_aware
function and modify its logic to use util_fits_cpu wherein it
returns the target CPU when the task fit. Also let's try our best
not to break schedtune by having boosted_task_util to be used in
the logic still.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>