power: qcom: smb-lib: Avoid getting stuck after switching to normal
For some reason, the MCU still refuses to raise the current after
switching to normal sometimes. Unfortunately, this shows up in the
driver as charging still since the MCU still sends enough current
to keep the charging state on but very small to have the SOC raise
at all.

To prevent this from happening, restart charging right after switching
to normal as replugging resolves this issue usually. This code
essentially allows the driver to do it on its own without needing
user intervention.

In addition to this, drop the status wait from 15 to 10 seconds as
it is unnecessarily long and it shouldn't take more than 10 seconds
for charging to begin after plugging.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>