clk: qcom: gcc-msm8998: Correct the HMSS_AHB_CLK_ENA clear bit operation
In k4.4, the address defined in the operation is the following:

+ #define GCC_APCS_CLOCK_BRANCH_ENA_VOTE				0x52004

wherein BIT(21) is bit AND NOT from the address.

This is currently not being done by Kholk's patch and has actually
bit OR BIT(21) into 0x52008 and not 0x52004. Not sure why that is,
but let's follow downstream for Linux to cease control of HMSS_AHB
and to let RPM manage the clock.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>