clk: qcom: gcc-msm8998: Partially revert gcc_usb30_master_clk patch
We actually don't need the .depends adjustment as the driver is
able to comply to this requirement without problems. Due to recent
changes, we now avoid setting the parent rate unncessarily in
order to passthrough the correct rate towards the clock source.

Return gcc_usb30_master_clk to the clock source.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>