Revert "regulator: rpm-smd: Reenable LEVEL and FLOOR_LEVEL for LDOA regulators"
This reverts commit b91b8c571a94384aad3f23722f0d473ad30d0dcb.

This actually isn't needed as the supported_regulator_types only
gets read for pinctrl enable and configuring optional init parameters.
The *voltage-level props get read by init regardless of it being
defined as supported as they are hard-coded to support both LDOA
and SMPS regulators and are used to determine the voltage index
to be updated.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>