sched: ucassist: Reduce input sleep timer to 1 second
With the external trigger configured at the KGSL IOCTL, having the
duration configured to 5 seconds causes the timeout to seldom occur
as the IOCTL gets called multiple times during that period even
during GPU idle. Hence, let's drop the duration of the timer to 1
second to save power in-between active input and GPU periods.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>