arm64: DT: msm8998: Restore USB3 core_base reg length
This isn't necessary as the usb-bam only uses registers from
0xa8f8000 + 0x300. There is no need to extend the register since
the code is hard-coded to only be limited to this range and also
that each register within this range are dedicated to specific
hardware functionalities.

Furthermore, we do not want HLOS to be accessing registers which
hardware does not expect to be accessed.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>