input: fingerprint: fpc1020: Ensure the sysfs notification reaches HAL
There are instances where even though the IRQ has been triggered,
the fingerprint HAL is not able to receive any data. This can be
caused by the system suspend happening too quickly after the
wake IRQ system wakeup was performed and before the IRQ handler is
called. __pm_wakeup_event only adds a wakeup event and does not call
on pm_system_wakeup() so it is not able to wakeup an already suspended
system.

To ensure that the system is awake when the wakelock is being held
and the sysfs is being notified, call pm_system_wakeup() in
the handler itself and unconditionally call on the wakeup event to
prevent the system from suspending if the IRQ occured closely prior
to system suspend.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>