power: port more changes from 4.4
Attempt to improve the charging situation, which still has 2 issues:
1) sometimes, the charge stops when switching from DASH to normal charge
2) sometimes, che charge stops due to charging adapter not being able to
provide a voltage >4.3V at the beginning of the charging process. This does
not happen with the original DASH charger, but it happens for example charging
the phone using USB hubs or computers.

This commit does not fix these issues, but it brings us closer to 4.4 which
does not have at least issue 1.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>