power: qcom: op_cg_uovp: Refactor op_cg_detect_normal and fix ceil_ichg_ua loop
We have layered op_cg_detect_normal current increase sequence
unnecessarily deep, perform inversion to make the code clearer
and easier to understand.

Also, we do not need to start from the bottom if we're looking for
the ceil_ichg_ua since we want the maximum current for that USB
type, reverse the loop to start from the highest current so that
we can bail right away when we find a match. We must ensure as well
that the APSD bit is updated, just incase it has been reran while
op_cg_uovp is running since we only check the bit at op_cg_uovp_enable.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>