power: qcom: smb-lib: Reset the charging GPIOs and remove not_charging WA
The not_charging workaround is not reliable and often fails. As such
it has been observed that replugging the dash charger does fix the
low current issue more reliably. Hence, like the watchdog rewrite
in the dash driver, let's perform the GPIO reset in check_dash_status
as well instead of the charging_en WA.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>