clk: qcom: gcc-msm8998: Allow GPLL main output consumers get rounded div rates
This is to allow the GPLL0 div legs get div-2 rounded rates from the
postdiv. Hardware sets FSM mode on these clocks so this should be
fine as it blocks any set rate on these postdiv clocks. Let's keep
the GCC children clocks to div-1 however since HW may have not set FSM
mode yet when probing the clocks, causing the postdivs to be probed
at div-2 which is wrong.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>