clk: qcom: gcc/mmcc/gpucc-msm8998: Configure proper Fabia PLL FMAXes
This is to properly scale the regulators powering these Fabia PLLs
according to specifications. This patch is based on sdm845, provided
that the platform also uses Fabia and Fabia Fixed PLLs.

Also remove the GPU_MX FMAX votes on the gfx3d_clk branch as this
was not carried over to sdm660 nor sdm845, and that this will now
be dependent on the PLL rate.

Reference:
https://github.com/LineageOS/android_kernel_xiaomi_sdm845/commit/d04d2ca99f3bf56baf54a183c9262aaebb027772

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>