arm64: DT: msm8998-regulator: Config msm8998 CPR controllers to use default step_quot
Tirupathi Reddy <tirupath@codeaurora.org>:
There is a chance for a CPR controller to use the run-time calibrated
step_quot of lower modes for higher modes. This can sometimes lead
to closed-loop voltage getting stuck at open-loop voltage. To avoid
this, configure APC0/1 and GFX CPR controllers in sdm660 to first use
the default step_quot and then later switch to the run-time calibrated
step_quot.

Reference:
https://github.com/LineageOS/android_kernel_nokia_sdm660/commit/cf855b2c65999f1e7ef1026b27710af3b11301a4

This has been applied to sdm660 which is based on the msm8998 platform
(and also sdm845, but the configuration was moved from Linux to secure
bootloader). Provided that both platforms use the same driver for
handling the said CPR controllers, there is a high probability that the
same can occur on msm8998.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>