power: qcom: op_cg_uovp: Evaluate UOVP immediately after changing current
Instead of waiting for another heartbeat interval, let's try to
correct the current as soon as the voltage settles to avoid staying
in an under/overvoltage state for unnecessarily long periods of time.
For this change, let's consolidate the UOVP evaluation within a single
function and implement a loop within op_cg_detect_uovp for the current
control and voltage evaluation.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>