devfreq: adreno_tz: Set the boost counter according to frame length
Using the MIN_BUSY as a condition for dropping the boost can cause
inconsistent performance. Instead, let's make use of the CEILING
macro which is used by the governor to determine whether to
unconditionally increase the frequency or not. In addition to that,
instead of resetting the counter immediately when the busy time
floor has been reached, let's instead gradually drop the counter
to avoid GPU frequency drops when there is an active GPU load that
is fluctuating due to CPU/IO bottleneck.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>