power: smb-lib: Fix charging issues due to incorrect intake voltage [2]
Apply 3 changes:
- try more times (3 -> 5), as this happens very quickly anyway (less that
a second usually);
- enable ICL override if needed: by default, when battery charges that
can provide more than 0.5A (excluding the original DASH charger) are used
(e.g. USB hubs with 1.5A ports, USB3 computers, wall chargers) the override
bit is not set, so we cannot read the software requested current. To prevent this,
when we noticed that the override bit is not set, set it -> read the current
value -> remove the bit again. This is a bit of a trick but it's the simpler
solution right now;
- increate timeout to 20ms which is not noticeable from the user perspective,
but it may help with certain chargers.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>