sched: ucassist: Allow configuration of timeout per trigger
We don't want to hold the input sleep state for long when triggered
by the GPU load. Allow us to configure the timeout per trigger by
saving the current timer timeout jiffies and comparing it to the new
one if it is later than the current.

Also increase the duration of input event timeout to 3 seconds.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>