power: qcom: op_cg_uovp: Avoid and remove unnecessary UOVP checks and configuration
Instead of using the UOVP state to decide whether to evaluate for
normal state or not, use the outcome of the non-hysteresis evaluate_uovp
instead as we now no longer update the state as soon as UOVP gets
evaluated due to the current control. Also remove the unnecessary
enable boolean.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>