power: qcom: op_cg_uovp: Prevent unmapped memory access
We may disable UOVP and clear all data including opdata->chg while
the op_heartbeat_work is running. This can cause unmapped memory
access if we clear the address while the op_check_charger_uovp is
running. Fix it.

Also bail early if we unplugged the USB when evaluating UOVP.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>