power: qcom: smb-lib: Monitor dash charging after switch to normal
There is a chance that normal charging fails on the dash charger
which causes it to completely stop or charge at a very low current.
To ensure this doesn't happen, let's monitor the charging after
update_dash_unplug_status and perform corrections to prevent this
from happening - or if it does happen, attempt to at least recover
and get the device charging again by allowing the SMB to fully
take over.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>