power: qcom: smb-lib: Rework check_dash_status to dash_to_normal_watchdog
We do not need to poll for the dash status anymore as our AICL adjustments
fix the dash status issue, hence we rename this function to watchdog
as it checks whether there is a problem with the power supply and perform
appropriate adjustments. Furthermore, make sure to notify the battery
of the power supply change, drop the unnecessary ICL override smblib writes
and also abort the function if we have already unplugged from the charger
as indicated by the power supply status.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>