clk: qcom: mdss-dsi-pll-10nm: Configure clk_regmap_div_mdss_ops
Like with the MMCC clocks, we want to make sure that the display PLLs
do not get latched on XO after power collapse. Let's configure div ops
for the DSI PLL branches to ensure that we always set the correct divider
configuration which the branch clocks use for selecting the parent clock.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>