sched: ucassist: Introduce runtime UCLAMP assist
This adds support to define taskgroup, global, and task-specific
UCLAMP values runtime configurations to allow granular control of
UCLAMP within the kernel.

Task-specific UCLAMP configuration:

This will allow us to specifically target important system and
kernel processes and define taskgroup-persistent UCLAMP values to
further reduce jitter without increasing system load significantly
as we only prioritize these defined targets.

This also makes it possible to target tasks beyond tasks that call
__setscheduler and adds support to also set UCLAMP values for processes
that call on kernel interfaces such as IOCTLs. In this case, since
our purpose is to reduce rendering jitter, we want to find and target
display and GPU processes that deal with processing graphical data.

For this purpose, we add hooks to kgsl_ioctl and mdss_fb IOCTL to
apply defined GPU and DISPLAY UCLAMP min values to boost these tasks
independent of their respective task groups as they can be placed
in task groups that are not top-app, which is what we usually boost.

Taskgroup and global UCLAMP configuration:

This is to reduce power consumption by globally restricting values
throughout UCLAMP. This is based on a timer-based system during runtime
by which the UCLAMP values will be restricted if timeout occurs, to be
reset by input triggers. On the other hand, if FB_BLANK occurs, we restrict
the UCLAMP values further, of which has a higher priority over runtime
restrictions.

With this, we can now finally control UCLAMP completely during runtime
based on renderthread and framebuffer events. Hence, let's completely
block UCLAMP writes from userspace to avoid overlapping and conflicting
values.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>