sched/uclamp: Add boosted UCLAMP flag
This is a separate flag that allows us to override uclamp_boosted which
is responsible for boosted task placement and is based on the uclamp.min
value. This is for cases wherein we want to keep a task group be biased
for placement within big cores in the scheduler even if there no longer
is a uclamp.min value.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>