sched: ucassist: Implement cross-driver task-specific UCLAMP assist support
This makes it possible to target tasks beyond tasks that call
__setscheduler and adds support to also set UCLAMP values for processes
that call on kernel interfaces such as IOCTLs. In this case, since
our purpose is to reduce rendering jitter, we want to find and target
display and GPU processes that deal with processing graphical data.

In this patch, we add hooks to kgsl_ioctl and mdss_fb IOCTL to apply
defined GPU and DISPLAY UCLAMP min values to boost these tasks
independent of their respective task groups as they can be placed
in task groups that are not top-app, which is what we usually boost.

Also cleanup the task list to only have display and GPU tasks.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>