sched: ucassist: Rework task filter UCFLAGs and add TRIGGER_UCFLAG
Let's rework how we filter tasks in order to better define how we filter
tasks according to the caller and the target's flag.

This allows us to drop the unnecessary trigger_input boolean and introduce
TRIGGER_UCFLAG to indicate whether a task is to trigger input and
also limit which callers are to trigger input as we do not want to
perform irq_work_queue on the display driver path nor at the scheduler
path.

Also sanitize unused code.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>